put classes into separate models.py file
