+++ readme.md
@@ -12,3 +12,4 @@ This is the Aliens example that comes with PyGame with a few enhancements
 ## Change Log
 * 2024-02-08 - added scaling based on device screen size.
 * 2024-02-09 - changed score display font size to adjust to game size
+* 2024-02-10 - put non-main program classes into a models.py file
