+++ settings.py
@@ -1,6 +1,11 @@
+from typing import ClassVar
+
 import pyautogui
+import pygame as pg
+
 
 class Settings:
+    SCORE: ClassVar[int] = 0
 
     def __init__(self):
         """Initialize the game's settings."""
@@ -13,7 +18,7 @@ class Settings:
         # round scaled width and height to multiple of 100
         game_width: int = int((device_width * self.screenPct // 100) * 100)
         game_height: int = int((device_height * self.screenPct // 100) * 100)
-        print(f'game width: {game_width}, height: {game_height}')
+        # print(f'game width: {game_width}, height: {game_height}')
 
         # self.scaleFactor = device_width / device_height * self.screenPct
         # print(f'scale factor = {self.scaleFactor}')
@@ -23,3 +28,5 @@ class Settings:
         self.near_bottom = game_height * 0.90
         self.score_board_y = game_height * 0.95
         self.score_font_size = int(device_height * 0.03)
+        self.SCREENRECT = pg.Rect(0, 0, self.screen_width, self.screen_height)
+
